reloc: new module that implements autoconf path expansion
In autoconf the installation layout is controlled by shell variables that
are expanded recursively.

For instance, we have a prefix variable and then this gets referenced in
sysconfdir by having the value '${prefix}/etc'

Earlier, the configure script did the expansion prior to putting these
values in config.h, now this is being pushed to the syslog-ng binary
so it can be relocated after compilation.

The new reloc module implements the shell-like substitution and
a PathResolver class which can be used with a Cache instance to resolve
autoconf-style directory layouts.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Tamas Pal <folti@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>