host-resolve: always return a new GSockAddr instance from resolve_hostname_to_sockaddr()
Instead of trying to _change_ an existing GSockAddr instance, always return
a new one. This makes the API easier to use, as well as makes
the implementation simpler. The original prototype probably dates back
to times when the afsocket driver stored ip/port parameters in an
existing GSockAddr instance and creating a new one was not good.

These times, we basically create the GSockAddr instance in the drivers
just to satisfy this requirement. So this patch simplifies call-sites
as well.

The only reason the input GSockAddr was used to determine the target address
family, but I introduced a new parameter for that.

Also the patch adds coverage for IPv6 in the testsuite.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>