host-resolve: refactor to smaller functions
This patch splits the monster functions in host-resolve to smaller,
more manageable parts.

This patch also changes the function prototypes to use
an internal, per-thread buffer instead of returning data in a buffer
passed as an argument. That saves a copy in the fast path and makes
the code easier to follow.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>